Add files via upload
Added "Tile" object which helps to track whether each location in the maze is the goal, a wall, etc, as well as the adjacent tiles of the maze. Have begun to work on the BFS, but it is glitchy at the moment. Will be doing more tomorrow, but need to stop working for now.